Refactor to `FollowWrapper` and `FollowButton`
This places the Bridget calls outside of the component, making it easier
to test and add stories for.
